feat(assignments-service): Move CourseAuth to CourseMember module and check for members
