+++ services/apps/assignments/src/course-member/course-auth.decorator.ts
